Move employee_project entity to correct namespace
