+++ src/main/java/de/szut/lf8_project/config/SampleDataCreator.java
@@ -2,8 +2,8 @@ package de.szut.lf8_project.config;
 
 import de.szut.lf8_project.employee.EmployeeEntity;
 import de.szut.lf8_project.employee.EmployeeRepository;
-import de.szut.lf8_project.employee.employee_project.EmployeeProjectEntity;
-import de.szut.lf8_project.employee.employee_project.EmployeeProjectRepository;
+import de.szut.lf8_project.employee_project.EmployeeProjectEntity;
+import de.szut.lf8_project.employee_project.EmployeeProjectRepository;
 import de.szut.lf8_project.project.ProjectEntity;
 import de.szut.lf8_project.project.ProjectRepository;
 import org.springframework.boot.ApplicationArguments;
