+++ src/main/java/de/szut/lf8_project/employee/EmployeeEntity.java
@@ -1,6 +1,6 @@
 package de.szut.lf8_project.employee;
 
-import de.szut.lf8_project.employee.employee_project.EmployeeProjectEntity;
+import de.szut.lf8_project.employee_project.EmployeeProjectEntity;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
