+++ src/main/java/de/szut/lf8_project/employee_project/EmployeeProjectRepository.java
@@ -1,4 +1,4 @@
-package de.szut.lf8_project.employee.employee_project;
+package de.szut.lf8_project.employee_project;
 
 import org.springframework.data.jpa.repository.JpaRepository;
