+++ src/test/java/de/szut/lf8_project/integrationtests/project/DeleteProjectTest.java
@@ -1,7 +1,7 @@
 package de.szut.lf8_project.integrationtests.project;
 
 import de.szut.lf8_project.employee.EmployeeEntity;
-import de.szut.lf8_project.employee.employee_project.EmployeeProjectEntity;
+import de.szut.lf8_project.employee_project.EmployeeProjectEntity;
 import de.szut.lf8_project.integrationtests.BaseIntegrationTest;
 import de.szut.lf8_project.project.ProjectEntity;
 import org.junit.jupiter.api.Test;
