+++ src/test/java/de/szut/lf8_project/testcontainers/AbstractIntegrationTest.java
@@ -1,7 +1,7 @@
 package de.szut.lf8_project.testcontainers;
 
 import de.szut.lf8_project.employee.EmployeeRepository;
-import de.szut.lf8_project.employee.employee_project.EmployeeProjectRepository;
+import de.szut.lf8_project.employee_project.EmployeeProjectRepository;
 import de.szut.lf8_project.project.ProjectRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
